DEV: Use new get_code in new_sequence.py and port tests
