+++ tests/test_core/test_sequence.py
@@ -2502,139 +2502,9 @@ def test_get_drawable(DATA_DIR):
         assert "(incomplete)" in trace.text
 
 
-@pytest.mark.parametrize("gc,seq", ((1, "TCCTGA"), (1, "ACGTAA---"), (2, "TCCAGG")))
-def test_has_terminal_stop_true(gc, seq):
-    gc = cogent3.get_code(gc)
-    seq = cogent3.make_seq(seq, moltype="dna")
-    assert seq.has_terminal_stop(gc=gc)
-
-
 @pytest.mark.parametrize(
-    "gc,seq", ((1, "TCCAGG"), (2, "TCCAAA"), (1, "CCTGA"), (2, "CCAGG"))
-)
-def test_has_terminal_stop_false(gc, seq):
-    gc = cogent3.get_code(gc)
-    seq = cogent3.make_seq(seq, moltype="dna")
-    assert not seq.has_terminal_stop(gc=gc)
-
-
-def test_has_terminal_stop_strict():
-    gc = cogent3.get_code(1)
-    seq = cogent3.make_seq("TCCAG", moltype="dna")
-    with pytest.raises(AlphabetError):
-        seq.has_terminal_stop(gc=gc, strict=True)
-
-
-@pytest.mark.parametrize(
-    "gc,seq",
-    (
-        (2, "TCCAGG"),
-        (1, "TAATGA"),
-        (1, "ACGTGA---"),
-        (1, "--AT-CTGA"),
-    ),
-)
-def test_trim_terminal_stop_true(gc, seq):
-    gc = cogent3.get_code(gc)
-    expect = re.sub("(TGA|AGG)(?=[-]*$)", "---" if "-" in seq else "", seq)
-
-    seq = cogent3.make_seq(seq, moltype="dna")
-    got = str(seq.trim_stop_codon(gc=gc))
-    assert got == expect
-
-
-@pytest.mark.parametrize("gc,seq", ((1, "T?CTGC"), (2, "TCCAAG")))
-def test_trim_terminal_stop_nostop(gc, seq):
-    gc = cogent3.get_code(gc)
-    seq = cogent3.make_seq(seq, moltype="dna")
-    got = seq.trim_stop_codon(gc=gc)
-    assert str(got) == str(seq)
-    # since there's no stop, we just return the same object
-    assert got is seq
-
-
-@pytest.mark.parametrize(
-    "gc,seq", ((1, "TCCAGG"), (2, "TCCAAA"), (1, "CCTGA"), (2, "CCAGG"))
-)
-def test_trim_terminal_stop_false(gc, seq):
-    gc = cogent3.get_code(gc)
-    seq = cogent3.make_seq(seq, moltype="dna")
-    assert str(seq.trim_stop_codon(gc=gc)) == str(seq)
-
-
-def test_trim_terminal_stop_strict():
-    gc = cogent3.get_code(1)
-    seq = cogent3.make_seq("TCCAG", moltype="dna")
-    with pytest.raises(AlphabetError):
-        seq.trim_stop_codon(gc=gc, strict=True)
-
-
-@pytest.mark.parametrize("cast", (int, numpy.int32, numpy.int64, numpy.uint8))
-def test_index_a_seq(cast):
-    seq = cogent3.make_seq("TCCAG", moltype="dna")
-    got = seq[cast(1)]
-    assert isinstance(got, Sequence)
-
-
-@pytest.mark.parametrize("cast", (float, numpy.float32))
-def test_index_a_seq_float_fail(cast):
-    seq = cogent3.make_seq("TCCAG", moltype="dna")
-    index = cast(1)
-    with pytest.raises(TypeError):
-        seq[index]
-
-
-@pytest.mark.parametrize("moltype", ("dna", "protein"))
-def test_same_moltype(moltype):
-    moltype = get_moltype(moltype)
-    seq = moltype.make_seq("TCCAG")
-    got = seq.to_moltype(moltype)
-    assert got is seq
-
-
-def test_gapped_by_map_segment_iter():
-    moltype = get_moltype("dna")
-    m, seq = moltype.make_seq("-TCC--AG").parse_out_gaps()
-    g = list(seq.gapped_by_map_segment_iter(m, allow_gaps=True, recode_gaps=False))
-    assert g == ["-", "TCC", "--", "AG"]
-
-
-@pytest.mark.parametrize("rev", (False, True))
-@pytest.mark.parametrize("sliced", (False, True))
-@pytest.mark.parametrize("start_stop", ((None, None), (3, 7)))
-def test_copied_parent_coordinates(sliced, rev, start_stop):
-    orig_name = "orig"
-    seq = DNA.make_seq("ACGGTGGGAC", name=orig_name)
-    start, stop = start_stop
-    start = start or 0
-    stop = stop or len(seq)
-    sl = slice(start, stop)
-    seq = seq[sl]
-    sliced_name = "sliced"
-    seq.name = sliced_name
-    assert seq.name == sliced_name
-    if rev:
-        seq = seq.rc()
-    copied = seq.copy(sliced=sliced)
-    assert copied.name == sliced_name
-    # matches original
-    assert copied.parent_coordinates() == seq.parent_coordinates()
-    # and expected -- the coordinate name always reflects the underlying sequence
-    assert copied.parent_coordinates() == (orig_name, start, stop, -1 if rev else 1)
-
-
-@pytest.mark.parametrize("rev", (False, True))
-def test_parent_coordinates(rev):
-    seq = DNA.make_seq("ACGGTGGGAC")
-    seq = seq[1:1]
-    if rev:
-        seq = seq.rc()
-    seq.name = "sliced"  # this assignment does not affect the
-    # note that when a sequence has zero length, the parent seqid is None
-    assert seq.parent_coordinates() == (None, 0, 0, 1)
-
-
-@pytest.mark.parametrize("cls", (Aligned, Sequence, SeqView, ArraySequence, str, bytes))
+    "cls", (Aligned, ArraySequence)
+)  # ported for Sequence, SeqView, str and bytes
 def test_coerce_to_seqview(cls):
     seq = "AC--GGTGGGAC"
     seqid = "seq1"
